Trying to run

+++ booklify-backend/src/main/java/com/booklify/domain/Payment.java
@@ -2,6 +2,7 @@ package com.booklify.domain;
 
 import com.booklify.domain.enums.PaymentStatus;
 import jakarta.persistence.*;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 //Lungela Nkunkuma 222491582
@@ -30,7 +31,7 @@ public class Payment {
     private PaymentStatus paymentStatus;
 
     @Column(nullable = false, precision = 10, scale = 2)
-    private double amountPaid;
+    private BigDecimal amountPaid;
 
     @Column(name = "payment_date", nullable = false, updatable = false)
     private LocalDateTime paymentDate = LocalDateTime.now();
@@ -70,7 +71,7 @@ public class Payment {
         return paymentStatus;
     }
 
-    public double getAmountPaid() {
+    public BigDecimal getAmountPaid() {
         return amountPaid;
     }
 
@@ -88,7 +89,7 @@ public class Payment {
         private Order order;
         private String paymentMethod;
         private PaymentStatus paymentStatus;
-        private double amountPaid;
+        private BigDecimal amountPaid;
         private LocalDateTime paymentDate;
 
         public Builder setPaymentId(Long paymentId) {
@@ -121,7 +122,7 @@ public class Payment {
             return this;
         }
 
-        public Builder setAmountPaid(double amountPaid) {
+        public Builder setAmountPaid(BigDecimal amountPaid) {
             this.amountPaid = amountPaid;
             return this;
         }
